+++ packages/riverpod/test/providers/future_provider/auto_dispose_future_provider_test.dart
@@ -26,14 +26,14 @@ void main() {
     expect(
       ref.state,
       const AsyncLoading<int>()
-          .copyWithPrevious(const AsyncData(0), isRefesh: false),
+          .copyWithPrevious(const AsyncData(0), isRefresh: false),
     );
     verifyOnly(
       listener,
       listener(
         const AsyncData(0),
         const AsyncLoading<int>()
-            .copyWithPrevious(const AsyncData(0), isRefesh: false),
+            .copyWithPrevious(const AsyncData(0), isRefresh: false),
       ),
     );
   });
@@ -110,7 +110,7 @@ void main() {
       listener(
         null,
         const AsyncLoading<int>()
-            .copyWithPrevious(const AsyncData(42), isRefesh: false),
+            .copyWithPrevious(const AsyncData(42), isRefresh: false),
       ),
     );